+++ echb/pages/admin.py
@@ -2,7 +2,7 @@ from django.contrib import admin
 
 from django_summernote.admin import SummernoteModelAdmin
 
-from .models import Page, Feedback, Video, VideoCategory, Subscriber, OldUser, MailingLog
+from .models import Page, Feedback, Subscriber, OldUser, MailingLog
 
 
 @admin.register(Page)
@@ -14,15 +14,6 @@ class AdminPage(SummernoteModelAdmin):
 @admin.register(Feedback)
 class FeedbackAdmin(admin.ModelAdmin):
     list_display = ['subject', 'email', 'name', 'created']
-
-
-@admin.register(Video)
-class VideoAdmin(admin.ModelAdmin):
-    list_display = ['title', 'youtube_link', 'accept_prayer_request', 'category', 'date', 'interesting_event']
-    list_filter = ('category',)
-
-
-admin.site.register(VideoCategory)
 
 
 @admin.register(Subscriber)
