+++ echb/pages/models.py
@@ -40,41 +40,6 @@ class Feedback(Audit):
         ordering = ['created']
 
 
-class VideoCategory(Audit):
-    title = models.CharField(max_length=100)
-    slug = models.SlugField()
-    icon = models.FileField(upload_to='video_categories', blank=True, null=True)
-
-    def __str__(self):
-        return self.title
-
-    class Meta:
-        verbose_name = 'видео категория'
-        verbose_name_plural = 'видео категории'
-
-
-class Video(Audit):
-    title = models.CharField(max_length=150)
-    youtube_link = models.CharField(max_length=200)
-    urgent_text = models.CharField(max_length=250, verbose_name="Текст срочного объявления", blank=True, null=True)
-    accept_prayer_request = models.BooleanField(default=False)
-    date = models.DateTimeField(default=timezone.now)
-    category = models.ForeignKey(VideoCategory, on_delete=models.CASCADE)
-    interesting_event = models.BooleanField(default=False)
-    text_for_request = models.CharField(max_length=200,
-                                        blank=True,
-                                        null=True,
-                                        help_text="Введите название для формы")
-
-    def __str__(self):
-        return self.title
-
-    class Meta:
-        verbose_name = 'видео'
-        verbose_name_plural = 'видео'
-        ordering = ['date']
-
-
 class Subscriber(Audit):
     email = models.EmailField(unique=True)
     activated = models.BooleanField(default=False)
