+++ echb/pages/urls.py
@@ -6,11 +6,6 @@ from . import views
 urlpatterns = [
     path('contacts/', views.ContactsFormView.as_view(), name='contacts'),
     path('contacts/thankyou/', views.ContactsThankYouView.as_view(), name='thankyou'),
-    path('about-us/online/', views.CurrentVideosListView.as_view(), name='online'),
-    path('about-us/online/thankyou/',
-         views.TemplateView.as_view(template_name="pages/prayer_thankyou.html"), name='video-detail-thankyou'),
-    path('about-us/online/<int:pk>/', views.VideoDetailView.as_view(), name='video-detail'),
-    path('interesting-events/', views.VideoListView.as_view(), name='interesting-videos'),
     path('subscriber/success/',
          TemplateView.as_view(template_name='pages/subscription_thankyou.html'),
          name='subscriber-success'),
