+++ echb/static/css/blocks/video-cat/video-cat.sass
@@ -4,12 +4,12 @@
     box-shadow: 0 2px 4px 0 rgba(0,0,0,.2)
     border-radius: 5px
     border: 1px solid $primary-black
-    display: flex
+    display: inline-flex
+    margin: 10px 0
+    width: 100%
 
 
     +element('list')
-        display: flex
-        flex-flow: row
         margin-top: 30px
 
     +element('title')
@@ -51,3 +51,4 @@
         +desktop()
             background-color: #2b5262
             display: block
+            width: 23%
